added authorization check for scale queries

+++ server/src/graphql/Scale.ts
@@ -1,3 +1,4 @@
+import { ApolloError } from "apollo-server-errors";
 import { objectType, extendType, nonNull, stringArg, intArg } from "nexus"
 import { ScaleModel } from "../models/scale"
 
@@ -5,7 +6,7 @@ export const Scale = objectType({
     name: "Scale",
     definition(t) {
         t.nonNull.id("id");
-        t.nonNull.string("email");
+        t.nonNull.string("userId");
         t.nonNull.string("goal");
         t.nonNull.int("sliderValue");
         t.nullable.string("chasingSuccessDescription");
@@ -32,7 +33,7 @@ export const CreateScale = extendType({
             type: "Scale",
             description: "Create a new scale", 
             args: {
-                email: nonNull(stringArg()),
+                userId: nonNull(stringArg()),
                 goal: nonNull(stringArg()),
                 sliderValue: intArg(),
                 chasingSuccessDescription: stringArg(),
@@ -56,16 +57,20 @@ export const UpdateScale = extendType({
             description: "Update a scale",
             args: {
                 id: nonNull(stringArg()),
-                email: stringArg(),
+                userId: nonNull(stringArg()),
                 goal: stringArg(),
                 sliderValue: intArg(),
                 chasingSuccessDescription: stringArg(),
                 avoidingFailureDescription: stringArg(),
             },
             resolve: async (_, args) => {
-                const { id, ...scale } = args
+                const scale = await ScaleModel.findById(args.id)
+                if(scale.userId != args.userId)
+                  throw new ApolloError("Unauthorized scale update", "UNAUTHORIZED_USER")
 
-                const response = await ScaleModel.findByIdAndUpdate(args.id, scale, {new: true})
+                const { id, ...updatedScale } = args
+
+                const response = await ScaleModel.findByIdAndUpdate(args.id, updatedScale, {new: true})
                 return response
             }
         })
@@ -79,9 +84,14 @@ export const DeleteScaleById = extendType({
             type: "Scale",
             description: "Delete a scale",
             args: {
-                id: nonNull(stringArg())
+                id: nonNull(stringArg()),
+                userId: nonNull(stringArg())
             },
             resolve: async (_, args) => {
+                const scale = await ScaleModel.findById(args.id)
+                if(scale.userId != args.userId)
+                  throw new ApolloError("Unauthorized scale delete", "UNAUTHORIZED_USER")
+
                 const response = await ScaleModel.findByIdAndDelete(args.id)
                 return response
             }
